Move logic to prepare label_infos to config
